+++ api/docs/docs.go
@@ -1545,6 +1545,10 @@ const docTemplate = `{
                     "type": "string",
                     "example": "Canyon"
                 },
+                "distance": {
+                    "type": "number",
+                    "example": 12500
+                },
                 "id": {
                     "type": "string",
                     "example": "550e8400-e29b-41d4-a716-446655440000"